Remove the 'data' key
The 'data' key was added as a measure to get around (perceived)
issues with the json schema system. As there is no reason for a
metada update message to contain both book and media update information
(they normally wouldn't, but this shouldn't be a constraint) there
is no need to define them separately, which means the "allOf" combinator
can be removed and the issue is no longer an issue.